move prettifyJsonString into global helpers
