+++ src/pages/namespace/JqPlayground/index.tsx
@@ -20,7 +20,7 @@ import Examples from "./Examples";
 import FormErrors from "~/componentsNext/FormErrors";
 import Input from "~/design/Input";
 import { JqQueryErrorSchema } from "~/api/jq/schema";
-import { prettifyJsonString } from "./utils";
+import { prettifyJsonString } from "~/util/helpers";
 import { useExecuteJQuery } from "~/api/jq/mutate/executeQuery";
 import { useTheme } from "~/util/store/theme";
 import { useTranslation } from "react-i18next";
